+++ datauploader/tasks.py
@@ -14,9 +14,227 @@ from celery import shared_task
 from django.conf import settings
 from open_humans.models import OpenHumansMember
 from datetime import datetime
+from fitbit.settings import rr
+from requests_respectful import (RespectfulRequester,
+                                 RequestsRespectfulRateLimitedError)
 
 # Set up logging.
 logger = logging.getLogger(__name__)
+
+
+# TODO: we don't need this, we should ideally re-queue a request that hits the exception.
+class RateLimitException(Exception):
+    """
+    An exception that is raised if we reach a request rate cap.
+    """
+
+    # TODO: add the source of the rate limit we hit for logging (fitit,
+    # internal global fitbit, internal user-specific fitbit)
+
+    pass
+
+@shared_task
+def fetch_fitbit_data(fitbit_member, access_token, fitbit_data=None):
+    '''
+    Fetches all of the fitbit data for a given user
+    '''
+    fitbit_urls = [
+        # Requires the 'settings' scope, which we haven't asked for
+        # {'name': 'devices', 'url': '/-/devices.json', 'period': None},
+
+        {'name': 'activities-overview',
+         'url': '/{user_id}/activities.json',
+         'period': None},
+        # interday timeline data
+        {'name': 'heart',
+         'url': '/{user_id}/activities/heart/date/{start_date}/{end_date}.json',
+         'period': 'month'},
+        # MPB 2016-12-12: Although docs allowed for 'year' for this endpoint,
+        # switched to 'month' bc/ req for full year started resulting in 504.
+        {'name': 'tracker-activity-calories',
+         'url': '/{user_id}/activities/tracker/activityCalories/date/{start_date}/{end_date}.json',
+         'period': 'month'},
+        {'name': 'tracker-calories',
+         'url': '/{user_id}/activities/tracker/calories/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-distance',
+         'url': '/{user_id}/activities/tracker/distance/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-elevation',
+         'url': '/{user_id}/activities/tracker/elevation/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-floors',
+         'url': '/{user_id}/activities/tracker/floors/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-minutes-fairly-active',
+         'url': '/{user_id}/activities/tracker/minutesFairlyActive/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-minutes-lightly-active',
+         'url': '/{user_id}/activities/tracker/minutesLightlyActive/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-minutes-sedentary',
+         'url': '/{user_id}/activities/tracker/minutesSedentary/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-minutes-very-active',
+         'url': '/{user_id}/activities/tracker/minutesVeryActive/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'tracker-steps',
+         'url': '/{user_id}/activities/tracker/steps/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'weight-log',
+         'url': '/{user_id}/body/log/weight/date/{start_date}/{end_date}.json',
+         'period': 'month'},
+        {'name': 'weight',
+         'url': '/{user_id}/body/weight/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'sleep-awakenings',
+         'url': '/{user_id}/sleep/awakeningsCount/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'sleep-efficiency',
+         'url': '/{user_id}/sleep/efficiency/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'sleep-minutes-after-wakeup',
+         'url': '/{user_id}/sleep/minutesAfterWakeup/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'sleep-minutes',
+         'url': '/{user_id}/sleep/minutesAsleep/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'awake-minutes',
+         'url': '/{user_id}/sleep/minutesAwake/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'minutes-to-sleep',
+         'url': '/{user_id}/sleep/minutesToFallAsleep/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'sleep-start-time',
+         'url': '/{user_id}/sleep/startTime/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+        {'name': 'time-in-bed',
+         'url': '/{user_id}/sleep/timeInBed/date/{start_date}/{end_date}.json',
+         'period': 'year'},
+    ]
+
+    # Set up user realm since rate limiting is per-user
+    print(fitbit_member.user)
+    user_realm = 'fitbit-{}'.format(fitbit_member.user.oh_id)
+    rr.register_realm(user_realm, max_requests=150, timespan=3600)
+    rr.update_realm(user_realm, max_requests=150, timespan=3600)
+
+    # Get initial information about user from Fitbit
+    headers = {'Authorization': "Bearer %s" % access_token}
+    query_result = requests.get('https://api.fitbit.com/1/user/-/profile.json', headers=headers).json()
+
+    # Store the user ID since it's used in all future queries
+    user_id = query_result['user']['encodedId']
+    member_since = query_result['user']['memberSince']
+    start_date = arrow.get(member_since, 'YYYY-MM-DD')
+
+    # Refresh token if the result is 401
+    # TODO: update this so it just checks the expired field.
+    # if query_result.status_code == 401:
+    #     fitbit_member._refresh_tokens()
+
+    # Reset data if user account ID has changed.
+    if 'profile' in fitbit_data:
+        if fitbit_data['profile']['encodedId'] != user_id:
+            logging.info(
+                'User ID changed from {} to {}. Resetting all data.'.format(
+                    fitbit_data['profile']['encodedId'], user_id))
+            fitbit_data = defaultdict(dict)
+        else:
+            logging.debug('User ID ({}) matches old data.'.format(user_id))
+
+    fitbit_data['profile'] = {
+        'averageDailySteps': query_result['user']['averageDailySteps'],
+        'encodedId': user_id,
+        'height': query_result['user']['height'],
+        'memberSince': member_since,
+        'strideLengthRunning': query_result['user']['strideLengthRunning'],
+        'strideLengthWalking': query_result['user']['strideLengthWalking'],
+        'weight': query_result['user']['weight'],
+    }
+
+    # Some block about if the period is none
+    for url in [u for u in fitbit_urls if u['period'] is None]:
+        if not user_id and 'profile' in fitbit_data:
+            user_id = fitbit_data['profile']['user']['encodedId']
+
+        # Build URL
+        fitbit_api_base_url = 'https://api.fitbit.com/1/user'
+        final_url = fitbit_api_base_url + url['url'].format(user_id=user_id)
+        # Fetch the data
+        try:
+            print(final_url)
+            r = rr.get(url=final_url, 
+                    headers=headers, 
+                    realms=["Fitbit", 'fitbit-{}'.format(fitbit_member.user.oh_id)])
+        except RequestsRespectfulRateLimitedError:
+            logging.info('Requests-respectful reports rate limit hit.')
+            raise RateLimitException()
+
+        fitbit_data[url['name']] = r
+
+    #Period year URLs
+    for url in [u for u in fitbit_urls if u['period'] == 'year']:
+        years = arrow.Arrow.range('year', start_date.floor('year'),
+                                arrow.get())
+        for year_date in years:
+            year = year_date.format('YYYY')
+
+            if year in fitbit_data[url['name']]:
+                logger.info('Skip retrieval {}: {}'.format(url['name'], year))
+                continue
+
+            logger.info('Retrieving %s: %s', url['name'], year)
+            # Build URL
+            fitbit_api_base_url = 'https://api.fitbit.com/1/user'
+            final_url = fitbit_api_base_url + url['url'].format(user_id=user_id,
+                                                                start_date=year_date.floor('year').format('YYYY-MM-DD'),
+                                                                end_date=year_date.ceil('year').format('YYYY-MM-DD'))
+            # Fetch the data
+            try:
+                print(final_url)
+                r = rr.get(url=final_url, 
+                        headers=headers, 
+                        realms=["Fitbit", 'fitbit-{}'.format(fitbit_member.user.oh_id)])
+            except RequestsRespectfulRateLimitedError:
+                logging.info('Requests-respectful reports rate limit hit.')
+                raise RateLimitException()
+
+            fitbit_data[url['name']][str(year)] = r
+
+    # Month period URLs/fetching
+    for url in [u for u in fitbit_urls if u['period'] == 'month']:
+        months = arrow.Arrow.range('month', start_date.floor('month'),
+                                arrow.get())
+        for month_date in months:
+            month = month_date.format('YYYY-MM')
+
+            if month in fitbit_data[url['name']]:
+                logger.info('Skip retrieval {}: {}'.format(url['name'], month))
+                continue
+
+            logger.info('Retrieving %s: %s', url['name'], month)
+            # Build URL
+            fitbit_api_base_url = 'https://api.fitbit.com/1/user'
+            final_url = fitbit_api_base_url + url['url'].format(user_id=user_id,
+                                                                start_date=month_date.floor('month').format('YYYY-MM-DD'),
+                                                                end_date=month_date.ceil('month').format('YYYY-MM-DD'))
+            # Fetch the data
+            try:
+                print(final_url)
+                r = rr.get(url=final_url, 
+                        headers=headers, 
+                        realms=["Fitbit", 'fitbit-{}'.format(fitbit_member.user.oh_id)])
+            except RequestsRespectfulRateLimitedError:
+                logging.info('Requests-respectful reports rate limit hit.')
+                print('Requests-respectful reports rate limit hit.')
+                print(r.text)
+                raise RateLimitException()
+
+            fitbit_data[url['name']][month] = r
+
+    print(fitbit_data)
+    return fitbit_data
 
 
 @shared_task
