+++ fitbit/settings.py
@@ -13,6 +13,7 @@ https://docs.djangoproject.com/en/2.0/ref/settings/
 import os
 import dj_database_url
 import logging
+from requests_respectful import RespectfulRequester
 
 logger = logging.getLogger(__name__)
 
@@ -53,6 +54,25 @@ OH_DELETE_FILES = OH_API_BASE + '/project/files/delete/'
 # Fitbit configuration
 FITBIT_CLIENT_ID=os.getenv('FITBIT_CLIENT_ID')
 FITBIT_CLIENT_SECRET=os.getenv('FITBIT_CLIENT_SECRET')
+
+if REMOTE is True:
+    from urllib.parse import urlparse
+    url_object = urlparse(os.getenv('REDIS_URL'))
+    logger.info('Connecting to redis at %s:%s',
+        url_object.hostname,
+        url_object.port)
+    RespectfulRequester.configure(
+        redis={
+            "host": url_object.hostname,
+            "port": url_object.port,
+            "password": url_object.password,
+            "database": 0
+        },
+        safety_threshold=5)
+
+# Requests Respectful (rate limiting, waiting)
+rr = RespectfulRequester()
+rr.register_realm("Fitbit", max_requests=3600, timespan=3600)
 
 if REMOTE is False:
     FITBIT_CALLBACK_URL = 'http://127.0.0.1:5000/complete/fitbit'
